Finish working script to run docker containers locally
The goal w. this commit is to make it easy to run the docker
containers locally and interact with them through the script
to speed up debugging.
The containers use a network together, and can be started
separately, which should make it easier to input sleep
time on the build server, when waiting for the db to startup.